Always use safe mode (remove not unsafe)
